Add command handlers for /start and /photo_id in commands.py
